requirements.txt
tensorflow
gradio
numpy
matplotlib
Pillow